+++ src/test/java/bridge/controller/MoveCommandTest.java
@@ -1,13 +1,15 @@
 package bridge.controller;
 
+import bridge.domain.MoveCommand;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.util.NoSuchElementException;
 
-import static bridge.controller.MoveCommand.DOWN;
-import static bridge.controller.MoveCommand.UP;
+import static bridge.domain.MoveCommand.DOWN;
+import static bridge.domain.MoveCommand.UP;
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class MoveCommandTest {
@@ -17,8 +19,8 @@ class MoveCommandTest {
         String upCommand = MoveCommand.parseValueToCommand(UP.getValue());
         String downCommand = MoveCommand.parseValueToCommand(DOWN.getValue());
 
-        Assertions.assertThat(upCommand).isEqualTo(UP.getCommand());
-        Assertions.assertThat(downCommand).isEqualTo(DOWN.getCommand());
+        assertThat(upCommand).isEqualTo(UP.getCommand());
+        assertThat(downCommand).isEqualTo(DOWN.getCommand());
     }
 
     @DisplayName("지정되지 않은 값을 넘겨주면 예외가 발생한다.")
